remove reference to TokenType enum

+++ contracts/prize-strategy/PeriodicPrizeStrategy.sol
@@ -275,7 +275,7 @@ abstract contract PeriodicPrizeStrategy is Initializable,
   /// @notice Awards collateral as sponsorship to a user
   /// @param user The user to whom the tickets are minted
   /// @param amount The amount of interest to mint as tickets.
-  function _awardToken(address user, uint256 amount, uint256 tokenIndex) internal {
+  function _awardToken(address user, uint256 amount, uint8 tokenIndex) internal {
     address[] memory _controlledTokens = prizePool.tokens();
     require(tokenIndex <= _controlledTokens.length, "PeriodicPrizeStrategy/award-invalid-token-index");
     address _token = _controlledTokens[tokenIndex];
